Add SR.check_dconf() method
Helper method of the base SR class, that checks the device-config object
of the calling SR for the existence of an arbitrary number of keys (provided
as input). The method is configured by default to raise an exception in the
event of any positive number of keys missing, but can return the set of missing
keys as well.

Signed-off-by: Kostas Ladopoulos <konstantinos.ladopoulos@citrix.com>