CA-171359: mount.cifs - Domain needs to be specified separately
The version of mount.cifs present in Dundee throws an error if username is
in the form '<domain>\<username>'. The domain, if it exists, needs to be
specified separately. The patch checks if the username supplied is in the
above form, and if yes, the domain is stripped and explicitly set in the
credentials file before calling mount.cifs.

Signed-off-by: Kostas Ladopoulos <konstantinos.ladopoulos@citrix.com>